pyproject.toml
[project]
name = "fb-alert-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "selenium>=4.39.0",
    "splinter>=0.21.0",
    "webdriver-manager>=4.0.2",
]
